Added clangtidy checks and reviewed entire project

+++ src/Morrigu/Renderer/ImageLoader.cpp
@@ -7,7 +7,7 @@ namespace MRG::ImageLoader
 {
 	unsigned char* loadFromFile(const char* filename, int* x, int* y, int* comp, int req_comp, bool flipped)
 	{
-		stbi_set_flip_vertically_on_load(flipped);
+		stbi_set_flip_vertically_on_load(static_cast<int>(flipped));
 		return stbi_load(filename, x, y, comp, req_comp);
 	}
 }  // namespace MRG::ImageLoader
